% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addClusters.R
\name{addClusters}
\alias{addClusters}
\title{Add Clusters}
\usage{
addClusters(seurat_ls, nn_lslsmat, sampleNames_v = "cohort")
}
\arguments{
\item{seurat_ls}{list of seurat objects}

\item{nn_lslsmat}{list of neighborhood lists with clusterNeighbors run}

\item{sampleNames_v}{vector of sample names to run. Mainly just use 'cohort'. See details}
}
\value{
seurat objects with new obj@meta.data columns.
}
\description{
Add neighborhood clusters to seurat objects
}
\details{
clusterNeighbors can be run on individual samples, but is most commonly run on the entire cohort of samples.
The default of this function will take the cohort clusters from clusterNeighbors and add them to each individual seurat object in seurat_ls
If an individual sample name from both seurat_ls and nn_lslsmat is also in sampleNames_v, then those individual clusters will also be added.
}
