% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ap_dp20Colors_v}
\alias{ap_dp20Colors_v}
\title{AP DP20 Colors}
\format{
\subsection{\code{ap_dp20Colors_v}}{

Named vector of length 32
\describe{
\item{values}{Hex Color}
\item{names}{Population Name}
}
}
}
\usage{
ap_dp20Colors_v
}
\description{
Colors used for Amanda DP20 populations
}
\keyword{datasets}
