% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildSeurat.R
\name{buildSeurat}
\alias{buildSeurat}
\title{Build Seurat Object using spatial data}
\usage{
buildSeurat(
  data_lsdf = NULL,
  intensity_df = NULL,
  spatial_df = NULL,
  meta_df = NULL,
  assayName_v = "mIHC",
  slideName_v
)
}
\arguments{
\item{data_lsdf}{list of intensity, spatial, and meta dfs output by reformatForSeurat}

\item{intensity_df}{data.frame of marker intensity values to add to seurat object (only used if data_lsdf is null)}

\item{spatial_df}{data.frame of spatial values to add to seurat object (only used if data_lsdf is null)}

\item{meta_df}{data.frame of metadata values to add to seurat object (only used if data_lsdf is null)}

\item{assayName_v}{character vector used to name assay when creating object}

\item{slideName_v}{character vector used to label the coordinate data}
}
\description{
Use the output of reformatForSeurat to build a spatial seurat object
}
\details{
Creates a seurat object with the given "count" data, metadata, and coordinates.
}
