% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkNeighbors.R
\name{checkNeighbors}
\alias{checkNeighbors}
\title{Check Neighbors}
\usage{
checkNeighbors(
  neighbors_nn,
  seurat_obj,
  cell_v,
  idCol_v = "OBJECTID",
  classCol_v = NULL,
  slideName_v,
  radius_v
)
}
\arguments{
\item{neighbors_nn}{neighborhood object output by findCellNeighbors}

\item{seurat_obj}{seurat object output by buildSeurat}

\item{cell_v}{character vector name of the ID of a cell to check}

\item{idCol_v}{name of column containing cell IDs}

\item{classCol_v}{name of column that contains cell identity classes (not required)}

\item{slideName_v}{name of the SpatialImage object}
}
\value{
Print a plot
}
\description{
Visual check of neighbor calculation
}
\details{
placeholder
}
