% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanCSV.R
\name{cleanCSV}
\alias{cleanCSV}
\title{Clean CSV data.frame}
\usage{
cleanCSV(df, grep_v = "^X$|^V1$", cols_v = NULL)
}
\arguments{
\item{df}{clean data.frame output by annotations script}

\item{grep_v}{character vector of regex pattern to grab columns that need removing. Use either this or cols_v.
This is default. set to NULL if you want to use cols_v.}

\item{cols_v}{character vector of columns to remove. Use either this or grep_v}
}
\value{
same data.frame as before, with specified columns removed
}
\description{
Current processing method causes these files to contain two extra columns that we don't want.
Use this to remove them
}
\details{
Current processing method causes the input files to contain two extra columns that we don't want.
Consider taking a detailed look at the processing and use data.table, write out tables without row.names, or
something else to eliminate this issue. Using either grep_v or cols_v, remove specified columns from data.frame
}
