% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterNeighbors.R
\name{clusterNeighbors}
\alias{clusterNeighbors}
\title{Cluster K Neighbors}
\usage{
clusterNeighbors(
  nn_lslsmat,
  sampleNames_v = "cohort",
  ks_lsv,
  seedClasses_v = "all"
)
}
\arguments{
\item{nn_lslsmat}{list of lists of neighbor matrices (count and pct) along with metadata}

\item{sampleNames_v}{vector of names in nn_lslsmat to run. Usually just run for cohort.}

\item{ks_lsv}{list of vectors of k's to test. One list element per sampleNames_v element, can have 1 or multiple k's to test per.}

\item{seedClasses_v}{vector of cell classes to use as seeds. Default is all, can be any of the 'class' values}
}
\value{
modified nn_lslsmat with columns for cluster results of each k in the 'pct' and 'meta' matrices
}
\description{
Cluster neighborhood percent matrices using provided k values
}
\details{
It takes a long time to run the clustering so only choose a few ks.
}
