% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixImageDimPlot.R
\name{fixImageDimPlot}
\alias{fixImageDimPlot}
\title{Fix output of ImageDimPlot}
\usage{
fixImageDimPlot(plot_gg)
}
\arguments{
\item{plot_gg}{a ggplot/patchwork object output by ImageDimPlot()}
}
\value{
object of classes "gtable", "gTree", "grob", and "gDesc". Note that print() will not display plot
must use plot() on the output.
}
\description{
Edit ggplot object to avoid white circles around points
}
\details{
ImageDimPlot uses pch 21 as default point. This renders fine on Rstudio and markdown,
but when printed to pdf, there is a white outline on each point, making the plot difficult to
view. Have to update the colour parameter from "white" to match the fill
}
