% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/halo_removeAnnotationDuplicates.R
\name{halo_removeAnnotationDuplicates}
\alias{halo_removeAnnotationDuplicates}
\title{Remove Annotation Duplicates}
\usage{
halo_removeAnnotationDuplicates(
  df,
  idCol_v = "OBJECTID",
  classCol_v = "class",
  mainClass_v = "Myeloid cells"
)
}
\arguments{
\item{df}{clean data.frame output by annotations script}

\item{idCol_v}{column that contains the cell IDs}

\item{classCol_v}{column that contains the cell type annotations}

\item{mainClass_v}{so far it appears that all duplicate IDs belong to one main class and an alternate class.
Not sure about this part entirely. Will this always be the case, or can there be multiple two-class combinations?}
}
\value{
list of two data.frames: (1) "clean" contains the input data.frame with duplicates removed.
(2) "dup" contains the duplicate IDs and their classes.
}
\description{
It's possible for the intensity and combination of markers in certain cells to cause that cell to be
assigned multiple labels. Must pick one.
}
\details{
Take the cleaned, annotated data and search for duplicated cell IDs. Grab the cell annotations for these cells
and split them. One annotation will be kept for subsequent analysis, the other will be removed and recorded.
}
