% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeCohort.R
\name{makeCohort}
\alias{makeCohort}
\title{Make Cohort}
\usage{
makeCohort(nn_lslsmat)
}
\arguments{
\item{nn_lslsmat}{list of list of neighborhood matrices}
}
\value{
list of cohort level objects
}
\description{
make cohort-level objects
}
