% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotClusterCompOfSamples.R
\name{plotClusterCompOfSamples}
\alias{plotClusterCompOfSamples}
\title{Plot Cluster Composition of Samples}
\usage{
plotClusterCompOfSamples(
  meta_dt,
  xVar_v,
  classCol_v = "class",
  plotColGrep_v = NULL,
  plotCol_v = NULL,
  colors_lsv,
  annot_lsv = NULL,
  annotColors_lsv
)
}
\arguments{
\item{meta_dt}{cohort metadata}

\item{xVar_v}{column in meta_dt to use for x-axis. Usually corresponds to sample, treatment, timepoint, etc.}

\item{classCol_v}{column name in meta_dt corresponding to cell classes}

\item{plotColGrep_v}{regex to grep columns from meta_dt to plot}

\item{plotCol_v}{column names in meta_dt to plot}

\item{annot_lsv}{list of annotations to add below the plot. e.g. annot_lsv = list("outName1" = "colName1", "outName2" = "colName2")}

\item{annotColors_lsv}{list of colors for annotations. list element names are same as annot_lsv names; list elements are named color vectors, names are values of annot_lsv colNames in meta_dt}

\item{colors_v}{list of vectors of colors for bars. Each lest element name should correspond to a plot column supplied by plotCol_v or plotColGrep_v.}
}
\value{
list of stacked bar charts - combined, count, pct and data tables of counts and pcts
}
\description{
Stacked bar chart of each samples' composition of cluster assignments.
}
\details{
can only use one of plotColGrep_v or plotCol_v
}
