% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotROI.R
\name{plotROI}
\alias{plotROI}
\title{Plot ROI}
\usage{
plotROI(
  seuratObj,
  fov_v,
  projName_v = "ap_dp20",
  colorBy_v = c("class", "RCN")
)
}
\arguments{
\item{seuratObj}{seurat object of mIHC data}

\item{fov_v}{name of the seurat object's field of view. Should be the sample name, can be found in \code{seuratObj@images}}

\item{projName_v}{name of project's color object from spatial package. Default to AP DP20}

\item{colorBy_v}{vector of columns to color by. Will grep for "^colorBy_v\link{i}$" as well as "^colorBy_v\link[=0-9]{i}*_cluster"}
}
\value{
list of ggplots
}
\description{
make a dim plot of the ROI colored by population or neighborhood.
}
\details{
Use 'class' to plot the populations. Use 'RCN' to plot all RCNs, specify RCN09 to just do one.
If column to plot is neither class nor RCN, auto-generate colors.
}
