% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSeedCellCompOfClusters.R
\name{plotSeedCellCompOfClusters}
\alias{plotSeedCellCompOfClusters}
\title{Plot Cell Composition of Clusters}
\usage{
plotSeedCellCompOfClusters(
  meta_dt,
  classCol_v = "class",
  plotColGrep_v = NULL,
  plotCol_v = NULL,
  colors_v
)
}
\arguments{
\item{meta_dt}{cohort metadata}

\item{classCol_v}{column name in meta_dt corresponding to cell classes}

\item{plotColGrep_v}{regex to grep columns from meta_dt to plot}

\item{plotCol_v}{column names in meta_dt to plot}

\item{colors_v}{vector of colors for bars}
}
\value{
list of stacked bar charts - combined, count, pct and data tables of counts and pcts
}
\description{
Stacked bar chart of each clusters' composition of seed cells.
}
\details{
can only use one of plotColGrep_v or plotCol_v
}
