% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processCounts.R
\name{processcounts}
\alias{processcounts}
\title{Process Counts}
\usage{
processcounts(countspath)
}
\arguments{
\item{countspath}{path to counts data}
}
\value{
data frame of counts
}
\description{
Read in count and gate data
}
