% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rcnColors_v}
\alias{rcnColors_v}
\title{RCN Colors}
\format{
\subsection{\code{rcnColors_v}}{

Named vector of length 21
\describe{
\item{values}{Hex Color}
\item{names}{RCN Name}
}
}
}
\usage{
rcnColors_v
}
\description{
Colors used for recurrent cellular neighborhoods (RCNs)
}
\keyword{datasets}
