% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeAnnotationDuplicates.R
\name{removeAnnotationDuplicates}
\alias{removeAnnotationDuplicates}
\title{Remove Annotation Duplicates}
\usage{
removeAnnotationDuplicates(
  df,
  idCol_v = "ObjectNumber",
  classCol_v = "class",
  keep_v = 2,
  verbose_v = F
)
}
\arguments{
\item{df}{clean data.frame output by annotations script}

\item{idCol_v}{column that contains the cell IDs}

\item{classCol_v}{column that contains the cell type annotations}

\item{keep_v}{value indicating which class to keep. Acceptable values are "first" for 1st class, "last" for last class,
or a numeric value (e.g. 2) indicating which to take.}

\item{verbose_v}{logical indicating whether or not to print messages.}
}
\value{
list of three data.frames: (1) "clean" contains the input data.frame with duplicates removed.
(2) "dup" contains the duplicate IDs and their classes. (3) is a summary of 2.
}
\description{
It's possible for the intensity and combination of markers in certain cells to cause that cell to be
assigned multiple labels. Must pick one.
}
\details{
Take the cleaned, annotated data and search for duplicated cell IDs. Grab the cell annotations for these cells
and split them. One annotation will be kept for subsequent analysis, the other will be removed and recorded.
}
