% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeNeighbors.R
\name{summarizeNeighbors}
\alias{summarizeNeighbors}
\title{Take the neighborhood output and determine which cell types they belong to}
\usage{
summarizeNeighbors(
  seurat_obj,
  neighbors_nn,
  classCol_v = "class",
  levels_v = NULL,
  metaCols_v = c("OBJECTID", "class", "CLASSIFIERLABEL")
)
}
\arguments{
\item{seurat_obj}{seurat object output by buildSeurat}

\item{neighbors_nn}{neighborhood object output by findCellNeighbors}

\item{classCol_v}{name of column that contains cell identity classes}

\item{levels_v}{vector of factor levels to assign to classCol_v. If NULL (default) will assign levels automatically}

\item{metaCols_v}{column names to extract for reference.}
}
\value{
list of data.frames. "mat" is counts while "pct" is percentages
}
\description{
Map the neighborhood$id values to cell types

To do
}
