% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangleNeighbors.R
\name{wrangleNeighbors}
\alias{wrangleNeighbors}
\title{Wranlge Neighbors}
\usage{
wrangleNeighbors(
  seurat_ls,
  neighbors_lsNN,
  sampleNames_v = NULL,
  overwrite_v = F,
  verbose_v = F,
  cellIDCol_v = "ObjectNumber",
  classCol_v = "class",
  metaCols_v = c(cellIDCol_v, classCol_v),
  levels_v = NULL,
  neighborDir_v,
  suffix_v = "_neighbors.xlsx"
)
}
\arguments{
\item{seurat_ls}{seurat object that has been wrangled for spatial mIHC (buildSeurat)}

\item{neighbors_lsNN}{list of neighborhood objects output by findCellNeighbors}

\item{sampleNames_v}{optional vector of names of samples in seurat_ls. If NULL (default), will use all samples}

\item{overwrite_v}{logical indicating if existing neighbor data should be overwritten (TRUE). If FALSE (default), existing data will be read in, not re-run.}

\item{verbose_v}{logical. If FALSE (default), no printing. If TRUE, will print for each sample whether it's being run or read in.}

\item{cellIDCol_v}{character vector indicating which column}

\item{classCol_v}{name of column that contains cell identity classes}

\item{metaCols_v}{column names to extract for reference}

\item{levels_v}{vector of factor levels to assign to classCol_v. If NULL (default), will assign levels automatically}

\item{neighborDir_v}{directory to hold neighbor excel outputs. Will check here to read in results first.}

\item{suffix_v}{file name suffix. Used to write and to search}
}
\value{
a list of either length(sampleNames_v) or length(seurat_ls). Each list element is a list of length 3, containing \code{mat}, \code{pct}, and \code{meta}
\enumerate{
\item \code{mat} = one row per cell, one column per potential neighbor class, cell values are the number of the column's cell class that are neighbors of that row's cell
\item \code{pct} = same as \code{mat}, but values are percentages of total neighbors for that cell
\item \code{meta} = same number of rows as \code{mat} and \code{pct} (one per cell), and whatever columns are specified in metaCols_v.
}
}
\description{
Either read in summarizeNeighbor results, or run it
}
