% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kClusterSweep.R
\name{kClusterSweep}
\alias{kClusterSweep}
\title{Test multiple k values for kmeans clustering}
\usage{
kClusterSweep(nn_pct, ks_v = 10)
}
\arguments{
\item{nn_pct}{neighbor matrix converted to percentages}

\item{ks_v}{either a single numeric value indicating to test k = 1 to k = ks_v;
or a range of values to test.}
}
\value{
to do
}
\description{
to do
}
\details{
to do
}
